214: Move the auth module to top level
