+++ backend/src/api/auth/mod.rs
@@ -1,11 +1,7 @@
 use crate::AppState;
 use axum::{routing::post, Router};
 
-pub mod error;
-pub mod google;
-pub mod jwt;
 pub(crate) mod login;
-pub mod middleware;
 
 // used for OpenAPI generation, maybe not picked up by the compiler as "used".
 #[allow(unused)]
@@ -15,5 +11,3 @@ pub use login::login;
 pub fn router() -> Router<AppState> {
     Router::<AppState>::new().route("/login", post(login::login))
 }
-
-// TODO: Refactor auth router, remove it
