+++ backend/src/open_api.rs
@@ -1,20 +1,10 @@
 use utoipa::OpenApi;
 
-//use crate::api::*;
-use crate::*;
+use crate::api::*;
 
 /// OpenAPI documentation for the RMoods server.
 ///
 /// All routes that should be documented in our interactive docs should be added here.
 #[derive(OpenApi)]
-#[openapi(paths(
-    // debug::lorem,
-    // debug::timeout,
-    // debug::subreddit_info,
-    // debug::post_comments,
-    // debug::user_info,
-    // debug::subreddit_posts,
-    // debug::user_posts,
-    auth::login::login
-))]
+#[openapi(paths(auth::login::login))]
 pub struct ApiDoc;
