+++ backend/src/websocket/peers.rs
@@ -1,4 +1,4 @@
-use crate::api::auth::google::{GoogleId, GoogleUserInfo};
+use crate::auth::google::{GoogleId, GoogleUserInfo};
 use crate::websocket::{ClientMessage, ConnectionId, WsUserId};
 use std::collections::HashMap;
 use tokio::sync::mpsc::Sender;
