+++ backend/src/websocket/socket_handler.rs
@@ -1,4 +1,4 @@
-use crate::api::auth::google::{GoogleId, GoogleUserInfo, JwtUserInfo};
+use crate::auth::google::{GoogleId, GoogleUserInfo, JwtUserInfo};
 use crate::websocket;
 use crate::websocket::{ClientMessage, SystemMessage};
 use axum::extract::ws::{Message, WebSocket};
